+++ docs/source/changelog.md
@@ -5,6 +5,7 @@
   * Rename `TableReference` to `ExternalTableReference`
   * Add support for `ExternalTableReference` to point to tables in external (i.e. not managed by `pipedag`) schemas. 
   * Remove support for `ExternalTableReference` that points to table in schema of current stage. I.e. `ExternalTableReference` can only point to tables in external schemas.
+- Fix bug where a `Task` that was declared lazy but provided a `Table` without a query string would always be cache valid.
 
 ## 0.6.9 (2024-01-24)
 - Update dependencies and remove some upper boundaries